pyproject.toml
[project]
name = "polyai-python"
version = "0.0.9"
authors = [
    { name="Akhlak Mahmood", email="akhlakm@gatech.edu" },
]
description = "Python API for locally hosted LLMs."
readme = "README.md"
requires-python = ">=3.6"

keywords = ["nlp", "development"]

classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
  "pylogg @ git+https://github.com/akhlakm/python-logg.git",
  "requests",
  "flask",
  "sshtunnel",
]

[project.urls]
"Homepage" = "https://github.com/akhlakm/polyai-python"
"Bug Tracker" = "https://github.com/akhlakm/polyai-python/issues"

[project.scripts]
polyai = "polyai.__main__:main"

[project.optional-dependencies]
# Install using `pip install .[server]`
server = [
    "sqlalchemy >= 2.0",
    "pandas",
    "psycopg2-binary",
    "sshtunnel",
    "python-dotenv",
    "pyenv-encrypt",
    "safetensors",
    "datasets",
    "sentencepiece",
    "transformers @ https://github.com/huggingface/transformers/archive/refs/heads/main.zip",
    "accelerate",
    "triton",
    "texttable",
    "toml",
    "tqdm",
    "numpy",
    "protobuf",
    "gunicorn",
    "spacy",
    "ninja >= 1.11.1",
    "websockets >= 11.0.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["polyai"]

